Changed the Prefixes class so that it uses the terminology from the OWL 2 specification. In particular, a prefix name now contains the terminating : character. That is, when one registers a prefix name with the Prefixes class, the prefix name must end in :.
This was done because there was a mismatch between HermiT and the OWL API. OWL API uses the OWL 2 terminology, and so all prefixes were returned with the terminating : character. HermiT's Prefixes class, however, expected prefix names to be without the : character; therefore, HermiT would abbreviate IRIs with a double colon. Of course, this could have been fixed at the interface between the two APIs; however, fixing the conceptual mismatch seemed more appropriate.

git-svn-id: svn://edison.comlab.ox.ac.uk/krr/2008/HermiT/trunk@1049 6b6e81c3-f1c2-4030-a0a9-a0569132be5c